+++ classes/Http/HttpErrorException.php
@@ -1,0 +1,4 @@
+<?php
+namespace AppZap\PHPFramework\Http;
+
+class HttpErrorException extends \Exception {}